Clean Up

+++ src/main/java/binnie/botany/craftgui/ControlColourOption.java
@@ -19,7 +19,7 @@ public class ControlColourOption extends ControlTextOption<IFlowerColour> {
 		this.addAttribute(Attribute.MouseOver);
 		CraftGUIUtil.moveWidget(this.textWidget, new IPoint(22, 0));
 		this.textWidget.setSize(this.textWidget.getSize().sub(new IPoint(24, 0)));
-		final int th = (int) CraftGUI.render.textHeight(this.textWidget.getValue(), this.textWidget.getSize().x());
+		final int th = CraftGUI.render.textHeight(this.textWidget.getValue(), this.textWidget.getSize().x());
 		final int height = Math.max(20, th + 6);
 		this.setSize(new IPoint(this.size().x(), height));
 		this.textWidget.setSize(new IPoint(this.textWidget.getSize().x(), height));
