+++ src/main/java/binnie/craftgui/minecraft/control/ControlPlayerInventory.java
@@ -13,7 +13,7 @@ public class ControlPlayerInventory extends Control {
 	private List<ControlSlot> slots;
 
 	public ControlPlayerInventory(final IWidget parent, final boolean wide) {
-		super(parent, (int) (parent.getSize().x() / 2) - (wide ? 110 : 81), (int) parent.getSize().y() - (wide ? 54 : 76) - 12, wide ? 220 : 162, wide ? 54 : 76);
+		super(parent, parent.getSize().x() / 2 - (wide ? 110 : 81), parent.getSize().y() - (wide ? 54 : 76) - 12, wide ? 220 : 162, wide ? 54 : 76);
 		this.slots = new ArrayList<>();
 		for (int row = 0; row < 3; ++row) {
 			for (int column = 0; column < 9; ++column) {
