+++ src/main/java/binnie/craftgui/minecraft/control/ControlSlotBase.java
@@ -41,7 +41,7 @@ public abstract class ControlSlotBase extends Control implements ITooltip {
 
 	@Override
 	public void onRenderBackground(int guiWidth, int guiHeight) {
-		final int size = (int) this.getSize().x();
+		final int size = this.getSize().x();
 		CraftGUI.render.texture(CraftGUITexture.Slot, this.getArea());
 		if (this.getSuperParent().getMousedOverWidget() == this) {
 			RenderUtil.drawGradientRect(new IArea(new IPoint(1, 1), this.getArea().size().sub(new IPoint(2, 2))), -2130706433, -2130706433);
